Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/floating-point-total-queue-size.sharedworker-expected.txt

CONSOLE MESSAGE: line 10: ReferenceError: Can't find variable: SharedWorker

FAIL floating-point-total-queue-size.js shared worker wrapper file ReferenceError: Can't find variable: SharedWorker
